Organize code a bit in aws_credential_providers
